Add Tests for case insensitive Lucene FST search.
